+++ actix-identity/src/middleware.rs
@@ -44,7 +44,6 @@ where
     S::Future: 'static,
     T: IdentityPolicy,
     B: MessageBody + 'static,
-    B::Error: Into<Error>,
 {
     type Response = ServiceResponse<EitherBody<B>>;
     type Error = Error;
@@ -80,7 +79,6 @@ where
     S::Future: 'static,
     T: IdentityPolicy,
     B: MessageBody + 'static,
-    B::Error: Into<Error>,
 {
     type Response = ServiceResponse<EitherBody<B>>;
     type Error = Error;