+++ actix-cors/CHANGES.md
@@ -1,7 +1,9 @@
 # Changes
 
 ## Unreleased - 2021-xx-xx
+* Update `actix-web` dependency to `4.0.0.beta-16`. [#217]
 
+[#217]: https://github.com/actix/actix-extras/pull/217
 
 ## 0.6.0-beta.7 - 2021-12-18
 * Update `actix-web` dependency to `4.0.0.beta-15`. [#216]